display/requirements.txt
Pillow>=10.0.0
spidev>=3.5
rpi-lgpio>=0.6
displayhatmini>=0.0.2
qrcode>=7.4.2
